New `--fast-forward` option, refactor, and changes
`--time-diff` has been changed to default to `0s` or messages must be submitted by the same character and with the same body at the same time to be deduplicated.

`--fast-forward` takes a timestamp (`YYYY-MM-DD HH:MM:SS`) to "fast-forward" to. This writes the left-most log-file to the output till it reaches the timestamp, drains all preceeding logs till timestamp, then continues as normal.